split test app into separate files
