+++ test/support/api.ex
@@ -1,0 +1,8 @@
+defmodule AshRbacTest.Api do
+  @moduledoc false
+  use Ash.Api
+
+  resources do
+    registry(AshRbacTest.Registry)
+  end
+end
