+++ test/support/calculation.ex
@@ -1,0 +1,9 @@
+defmodule AshRbacTest.Calculation do
+  @moduledoc false
+  use Ash.Calculation
+
+  @impl true
+  def calculate(records, _, _) do
+    Enum.map(records, fn _ -> 1 end)
+  end
+end
